feat(hyper-boring): upgrade to hyper v1
