+++ Cargo.toml
@@ -8,7 +8,7 @@ members = [
 resolver = "2"
 
 [workspace.package]
-version = "4.6.0"
+version = "4.7.0-rc.1"
 repository = "https://github.com/cloudflare/boring"
 edition = "2021"
 
@@ -19,9 +19,9 @@ tag-prefix = ""
 publish = false
 
 [workspace.dependencies]
-boring-sys = { version = "4.6.0", path = "./boring-sys" }
-boring = { version = "4.6.0", path = "./boring" }
-tokio-boring = { version = "4.6.0", path = "./tokio-boring" }
+boring-sys = { version = "4.7.0-rc.1", path = "./boring-sys" }
+boring = { version = "4.7.0-rc.1", path = "./boring" }
+tokio-boring = { version = "4.7.0-rc.1", path = "./tokio-boring" }
 
 bindgen = { version = "0.68.1", default-features = false, features = ["runtime"] }
 cmake = "0.1.18"
@@ -36,9 +36,12 @@ futures = "0.3"
 tokio = "1"
 anyhow = "1"
 antidote = "1.0.0"
-http = "0.2"
-hyper = { version = "0.14", default-features = false }
+http = "1.0"
+hyper = { version = "1.0", default-features = false }
+hyper-util = "0.1"
+http-body-util = "0.1.0"
 linked_hash_set = "0.1"
 once_cell = "1.0"
 tower = "0.4"
 tower-layer = "0.3"
+tower-service = "0.3"
